backend-python/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
mysqlclient==2.2.4
python-dotenv==1.0.1
marshmallow==3.20.2
auth0-python==4.2.0
python-jose==3.3.0
flask-cors==4.0.0 